Added a missing modular reduction
The missing modular reduction may cause errors to be incorrectly detected and reported by the unit tests executable. The problem seemingly only arises for certain compilers, which explains why this problem was not previously detected.

For example, when the unit test executable is compiled with gcc under Ubuntu 18.04 LTS, no error is detected, whilst an error is detected when the same executable is compiled with gcc under Ubuntu 20.04 LTS.