Swapped HKZ for LLL in the reduction of L^\alpha.
Swapped HKZ for LLL in the reduction of L^\alpha, as using HKZ in rare events causes fpLLL to hang, and as the choice is not very significant.

In the longer term, it would be better to find the reason for why the function hangs, and/or to implement lattice reduction for small-dimensional lattices in Qunundrum itself to facilitate tracking down these kind of bugs.